refactor: rename for better understanding
